Restore working with pkgsrc 3.3.0 binutils
Up to and including release 3.3.0 of MINIX pkgsrc, the as assembler of the
package follow the AT&T convention to use / as comment starter, unlike
the behaviour of in-tree NetBSD as.
As of pkgsrc 3.4.0, the MINIX-specific binutils package has been changed
and now behaves the same as in-tree.

Unfortunately, commit 2aa0776 did not clearly anticipate the fact that
MINIX 3.4.0 can still be build from MINIX using packages 3.3.0, and
removed the AFLAGS+= -Wa,--divide helpers. Here they are restored.

Change-Id: I0001b5592e3ba40e6341ca46fc0f04b47468021a